[FIX] account_payment_pro_receiptbook: Fix the setting of l10n_latam_document_type_id when having receitbook.
The fix ensures that when using a receipt, the `l10n_latam_document_type_id` field is set by inheriting
its compute method instead of setting it only at the moment of posting.

Previously, if the field was set only in the action post when using receitbook,
any recomputation could reset it to False. By overriding the compute method,
the correct document type is assigned when using a receipt, preventing unintended resets.

Part-of: ingadhoc/account-payment#633