[FIX] account_payment_pro_receiptbook: Ensure to set l10n_latam_document_type_id when using receiptbook in payment.
Previously, we were computing l10n_latam_document_type_id
instead of assigning it in the post method (as in previous versions).
From v18, it was only being set in the compute method.

Then, due to an issue detected in this commit,
https://github.com/ingadhoc/account-payment/commit/8a6ff0564d3526ec8ead24c90a8e53267d038f6a
we attempted to fix a case where the recomputation left it unassigned,
both in the post method and the recompute, affecting receipts.(only in v18)

closes ingadhoc/account-payment#663

Signed-off-by: Filoquin adhoc <maq@adhoc.com.ar>