[FIX] account_payment_pro: Fix when calculating to_pay_move_line_ids when using direct debit mandate
When using direct debit, the invoice to be linked to the payment is already explicitly defined.
The standard computation of to_pay_move_line_ids should not be applied in this case, as it may incorrectly associate the payment with older unpaid invoices from the same vendor that are not related to the direct debit.

This fix bypasses the default computation for direct debit flows, ensuring that only the intended invoice is linked to the payment.

closes ingadhoc/account-payment#751

Signed-off-by: Filoquin adhoc <maq@adhoc.com.ar>